+++ Runtime/ConsentRequestParameters.cs
@@ -26,6 +26,9 @@ namespace CAS.UserConsent
         internal bool withRequestTrackingTransparency = false;
 
         [SerializeField]
+        private string trackingUsageDescription;
+
+        [SerializeField]
         internal UserConsentUI uiPrefab;
 
         [SerializeField]
@@ -157,6 +160,11 @@ namespace CAS.UserConsent
         {
             resetStatus = 2;
             return this;
+        }
+
+        public string defaultIOSTrakingUsageDescription
+        {
+            get { return withRequestTrackingTransparency ? trackingUsageDescription : null; }
         }
         #endregion